Fix: Migration LibGDX - Correction des erreurs de compilation

+++ core/src/main/java/com/main/entities/units/Soldier.java
@@ -4,7 +4,7 @@ import com.main.entities.Unit;
 
 public class Soldier extends Unit {
  
-    public Soldier (String filePath, int posX, int posY) {
+    public Soldier (String filePath, float posX, float posY) {
         super(filePath, posX, posY);
     }
 }
